Support receiving fragmented STOMP frames
Support receiving STOMP frame split into multiple Websocket
messages, since this is the current behavior of Stomp.js and
Dart STOMP library when frames are bigger than 16 Kbytes.

This commit also introduces a new maxFrameSize parameter
(default 1 Mbytes) in order to allow to configure the maximum
size of a whole STOMP frame (fragmented or not). This parameter
can be set globally or by STOMP end point.

This parameter is also used to define the size of the fragmented
STOMP frame buffer.

Issue: SPR-11527