Make @ResponseBody method return type available for message converters
This commit introduces ParametrizedTypeResponseBodyAdvice, useful
for keeping parametrized type information when serializing objects.

It uses ParameterizedTypeValue wrapper to retain the parametrized type
information available in the @ResponseBody method return type and
pass it to the HttpMessageConverter that implements the new
ParametrizedTypeAware marker interface.

AbstractJackson2HttpMessageConverter uses this new mechanism to make
@ResponseBody method return type available for Jackson type resolution
instead of just letting Jackson trying to guess the type to use from
the object to serialize.

Issue: SPR-12811