Revisit nullability annotations
This commit introduces 4 following changes.

1) It adds a new Spring @NonNull annotation which allows to apply
@NonNullApi semantic on a specific type, like @Nullable does.
Combined with @Nullable, it allows partial null-safety support when
package granularity is too broad.

2) @Nullable and @NonNull can apply to ElementType.TYPE_USE in order
to be used on generic type arguments (SPR-15942).

3) Annotations does not apply to ElementType.TYPE_PARAMETER anymore
since it is not supported yet (applicability for such use case is
controversial and need to be discussed).

4) @NonNullApi does not apply to ElementType.FIELD anymore since in a
lot of use cases (private, protected) it is not part for the public API
+ its usage should remain opt-in. Instead, non-nullable fields should be
annotated with @NonNull.

Issue: SPR-15756