Fix RSocketRequester API for requests without payload
This commit makes it possible to send requests without
requiring to call data(Mono.empty()).

It clarifies the semantics by renaming ResponseSpec
to InteractionSpec in order to leverage more directly RSocket
vocabulary and make things more consistent now that RequestSpec
extends it.

It also improves the Consumer variant of metadata in order to prevent
calling data in the callback (only metadata is now accessible).

Closes gh-23649